requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt>=4.0.1,<4.1.0
pydantic
python-dotenv
pydantic-settings
email-validator